[Silver III] Title: 주유소, Time: 376 ms, Memory: 37548 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -15,7 +15,7 @@ class Main {
         }
 
         int minCost = 1000000000;
-        int answer = 0;
+        long answer = 0;
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i < n - 1; i++) {
             int cost = Integer.valueOf(st.nextToken());
@@ -24,7 +24,7 @@ class Main {
                 minCost = cost;
             }
 
-            answer += minCost * distances[i];
+            answer += (long) minCost * distances[i];
         }
 
         System.out.print(answer);
